♻️ TS13 : MENUORDER > Migrate MenuOrder to TypeScript.

+++ src/components/pages/order/MainOrder/MenuOrder/MenuOrder.tsx
@@ -1,17 +1,17 @@
 import styled from "styled-components";
-import { Card } from "../../../../ui/Card/Card";
-import { formatPrice } from "../../../../../utils/maths";
-import { theme } from "../../../../../theme";
-import { useOrderContext } from "../../../../../contexts/OrderContext";
+import { Card } from "@/components/ui/Card/Card";
+import { formatPrice } from "@/utils/maths";
+import { theme } from "@/theme/theme";
+import { useOrderContext } from "@/contexts/OrderContext";
 import EmptyMenu from "./EmptyMenu";
 import { checkCardIsSelected } from "./helpers/checkCardIsSelected";
-import { DEFAULT_IMG, UNAVAILABLE_PRODUCT_IMG } from "../../../../../constants/product";
-import { findObjectById, isEmptyArray } from "../../../../../utils/array";
+import { DEFAULT_IMG, UNAVAILABLE_PRODUCT_IMG } from "@/constants/product";
+import { findObjectById, isEmptyArray } from "@/utils/array";
 import { Loader } from "./Loader";
 import { CSSTransition, TransitionGroup } from "react-transition-group";
-import { ProductCardAnimation } from "../../../../../theme/animations";
-import { convertStringToBoolean } from "../../../../../utils/string";
-
+import { ProductCardAnimation } from "@/theme/animations";
+import { convertStringToBoolean } from "@/utils/string";
+import { BasketProduct } from "@/types/Product";
 
 
 export const MenuOrder = () => {
@@ -19,32 +19,44 @@ export const MenuOrder = () => {
     const isAdminMode = adminMode.isAdminMode
     const cardSelected = productSelected.id
 
-    const onDelete = (productId, event) => {
+    const onDelete = (productId: string, event: React.MouseEvent<HTMLSpanElement, MouseEvent>) => {
         event.stopPropagation()
         handleDeleteBasketProduct(productId, userId)
         handleDeleteProduct(productId, userId)
     }
 
-    const onClick = (id) => {
+    const onClick = (id: string, isAdminMode: boolean) => {
+        if (!products) return
+
         const productSelected = findObjectById(id, products)
-        handleSelectProduct(productSelected)
+        productSelected && isAdminMode && handleSelectProduct(productSelected)
     }
 
-    const addToBasket = (id, event) => {
+    const addToBasket = (id: string, event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
         event.stopPropagation()
+        if (!products) return
+
         const productToAdd = findObjectById(id, products)
-        handleAddBasketProduct(productToAdd, userId)
+        productToAdd && handleAddBasketProduct(productToAdd, userId)
     }
 
-    const removeFromBasket = (id, event) => {
+    const removeFromBasket = (id: string, event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
         event.stopPropagation()
         decrementQuantityProduct(id)
 
         const removedProduct = findObjectById(id, basketProducts)
 
-        if (removedProduct.quantity < 2) {
+
+        if (removedProduct && removedProduct.quantity < 2) {
             handleDeleteBasketProduct(removedProduct.id, userId)
         }
+    }
+
+    const getProductQuantity = (productId: string, basketProducts: BasketProduct[]) => {
+        const basketProduct = basketProducts.find(basketProduct => basketProduct.id === productId)
+        const productQuantity = basketProduct?.quantity ?? 0
+
+        return productQuantity
     }
 
     const isLoading = products === undefined
@@ -73,14 +85,14 @@ export const MenuOrder = () => {
                                         hasDeleteButton={isAdminMode}
                                         onDelete={(event) => onDelete(id, event)}
                                         isHoverable={isAdminMode}
-                                        onClick={isAdminMode ? (() => onClick(id)) : undefined}
+                                        onClick={() => onClick(id, isAdminMode)}
                                         isSelected={checkCardIsSelected(id, cardSelected)}
                                         onAdd={(event) => addToBasket(id, event)}
                                         onRemove={(event) => removeFromBasket(id, event)}
                                         isPublicised={convertStringToBoolean(isPublicised)}
                                         isUnavailable={convertStringToBoolean(isAvailable) === false}
                                         unavailableImage={UNAVAILABLE_PRODUCT_IMG}
-                                        quantity={basketProducts.find(basketProduct => basketProduct.id === id) ? basketProducts.find(basketProduct => basketProduct.id === id).quantity : 0}
+                                        quantity={getProductQuantity(id, basketProducts)}
                                     />
                                 </CSSTransition>
                             )
